Fixed errors when running lock and extended wait
The timeout is only needed in emergencies. Putting it too low will make it so that if the main thread is not updating fast enough, it will use more cpu. This could make the main thread hang even more.

Also moved thread initialization into the IO::initialize function instead of the IO constructor.